Refactor data load for neural network
